refactor(frontend): standardize component file structure, migrate 10 components (#898)
Documents the folder convention in DESIGN_SYSTEM.md (ComponentName/
with index.tsx + colocated ComponentName.test.tsx / .stories.tsx,
no __tests__/ subfolder or separate stylesheet since styling is
Tailwind-only) and migrates 10 components to it:

- components/ui/Badge, EmptyState, StatCard, Checkbox, Switch,
  VerificationBadge, SearchBar
- components/OfflineIndicator
- components/analytics/CarbonImpactCard
- components/qr/QRGenerator

Import paths via the `@/components/...` alias are unaffected since
`index.tsx` resolves the same as the old flat `ComponentName.tsx`.
Relative imports inside moved files were adjusted for the new nesting
depth. Remaining components keep their current layout — this migrates
the 10 requested, not a wholesale rewrite.
