+++ frontend/src/components/OfflineIndicator/OfflineIndicator.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi } from 'vitest'
 import { render, screen } from '@testing-library/react'
-import { OfflineIndicator } from '@/components/OfflineIndicator'
+import { OfflineIndicator } from './index'
 
 // Mock the useOnlineStatus hook
 vi.mock('@/hooks/useOnlineStatus', () => ({
