+++ frontend/src/components/OfflineIndicator/index.tsx
