+++ frontend/src/components/qr/QRGenerator/index.tsx
@@ -1,6 +1,6 @@
 import { useRef } from 'react'
-import { Button } from '../ui/Button'
-import { Card } from '../ui/Card'
+import { Button } from '../../ui/Button'
+import { Card } from '../../ui/Card'
 
 interface QRGeneratorProps {
   wasteId: string
