+++ frontend/src/components/ui/Checkbox/Checkbox.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Checkbox } from './Checkbox'
+import { Checkbox } from './index'
 
 const meta: Meta<typeof Checkbox> = {
   title: 'Design System/Checkbox',
