+++ frontend/src/components/ui/EmptyState/index.tsx
@@ -1,5 +1,5 @@
 import { LucideIcon } from 'lucide-react'
-import { Button } from './Button'
+import { Button } from '../Button'
 import { cn } from '@/lib/utils'
 
 export interface EmptyStateProps {
