+++ frontend/src/components/ui/SearchBar/index.tsx
