+++ frontend/src/components/ui/StatCard/index.tsx
